docs: update tagging rules
As the project grows, tags like feature, fix, cleanup, and chore
have become outdated and less descriptive. To improve
maintainability and clarity, we are adopting a new tagging system
based on the latest guidelines. This change will make our project
easier to maintain by ensuring that commit messages better reflect
the purpose and scope of changes.

Signed-off-by: Jiamu Sun <barroit@linux.com>